requirements.txt
azure-storage-blob
streamlit
python-dotenv
psycopg2-binary

